[clean] Separate utility functions into a file
